requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
spacy>=3.5.0
nltk>=3.8.0

# API and Web Interface
fastapi>=0.95.0
streamlit>=1.22.0
uvicorn>=0.22.0

# Text-to-Speech
gTTS>=2.3.0
azure-cognitiveservices-speech>=1.28.0

# Video Generation
moviepy>=1.0.3
opencv-python>=4.7.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 